runtime: storage: Add random distribution bucket selection.
